Setting up buffers (index, vertexarray, vertexbuffer)

+++ core/maths/constants.h
@@ -6,9 +6,10 @@
 #define CONSTANTS_H
 
 #include <numbers>
+#include <glad/glad.h>
 
 namespace sparky {
-    typedef double real;
+    typedef GLfloat real;
 
     inline real toRadian(const real degree) {
         constexpr real pi_over_180 = std::numbers::pi/ 180; // For optimization
